AnomlyDetection/backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
numpy
opencv-python
Pillow